feat: 호선 정보 constants 파일로 분리
